+++ DTCD-WorkspacePanel/src/components/WorkspaceElementIcon.vue
@@ -1,16 +1,7 @@
 <template>
-  <div class="icon-wrapper">
-    <svg class="icon-back" :width="iconSize" :height="iconSize" :viewBox="viewBox" fill="none">
-      <rect width="100%" height="100%" :rx="iconRadius" :fill="backFill"/>
-      <defs>
-        <linearGradient id="g" :x1="iconSize" :y1="iconSize" x2="0" y2="0" gradientUnits="userSpaceOnUse">
-          <stop :stop-color="secondColor"/>
-          <stop :stop-color="mainColor" offset="1"/>
-        </linearGradient>
-      </defs>
-    </svg>
-    <div v-if="isFolder" class="icon-content" v-html="folderIcon"/>
-    <span v-else :class="['icon-content FontIcon size_3xl', iconClass]"></span>
+  <div class="IconWrapper" :style="iconStyle">
+    <div v-if="isFolder" class="Icon" v-html="folderIcon"/>
+    <span v-else :class="['FontIcon size_3xl Icon', iconClass]"></span>
   </div>
 </template>
 
@@ -29,12 +20,14 @@ export default {
     icon: { type: Number, default: 0 },
   },
   computed: {
-    viewBox() {
-      return `0 0 ${this.iconSize} ${this.iconSize}`;
+    background() {
+      return this.isFolder ? 'none' : `linear-gradient(to bottom right, ${this.mainColor}, ${this.secondColor})`;
     },
 
-    backFill() {
-      return this.isFolder ? 'none' : 'url(#g)';
+    iconStyle() {
+      const width = `${this.iconSize}px`;
+      const borderRadius = `${this.iconRadius}px`;
+      return { borderRadius, width, height: width, background: this.background };
     },
 
     iconClass() {
@@ -54,32 +47,25 @@ export default {
     },
 
     mainColor() {
-      return elementColors[this.colors = [0]];
+      return elementColors[this.colors[0]] || elementColors[0];
     },
 
     secondColor() {
       return this.colors.length < 2
         ? this.mainColor
-        : elementColors[this.colors[1]];
+        : elementColors[this.colors[1]] || elementColors[0];
     },
   },
 };
 </script>
 
 <style lang="sass" scoped>
-.icon-wrapper
-  position: relative
+.IconWrapper
+  display: flex
+  align-items: center
+  justify-content: center
 
-  .FontIcon
-    color: var(--background_main)
-
-  .icon-back
-    display: block
-
-  .icon-content
+  .Icon
     display: flex
-    position: absolute
-    top: 50%
-    left: 50%
-    transform: translate(-50%, -50%)
+    color: var(--background_main)
 </style>