+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - form settings for element
 
+## Fixed
+
+- workspace element colors are not applied or applied incorrectly
+
 ## [0.5.0]
 
 ### Added